requirements.txt
aiofiles==23.2.1
aiogram==3.6.0
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
attrs==23.2.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
frozenlist==1.4.1
google-auth==2.29.0
google-auth-oauthlib==1.2.0
gspread==6.1.0
idna==3.7
magic-filter==1.0.12
multidict==6.0.5
oauthlib==3.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.7.1
pydantic_core==2.18.2
python-dotenv==1.0.1
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
StrEnum==0.4.15
typing_extensions==4.11.0
urllib3==2.2.1
yarl==1.9.4
